requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
asn1crypto==1.3.0
astor==0.8.0
astroid @ file:///C:/ci/astroid_1592481955828/work
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.1
bitarray==1.2.2
bkcharts==0.2
bleach==3.1.5
blinker==1.4
bokeh==2.1.0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.20
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1592842333140/work
decorator==4.4.2
defusedxml==0.6.0
distributed @ file:///C:/ci/distributed_1592842563762/work
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
fsspec==0.7.4
gast==0.2.2
gevent==1.4.0
glob2==0.7
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.27.2
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==5.16.1
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///C:/ci/ipython_1592330400575/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1592846084747/work
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
nose==1.3.7
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.5
numpydoc==1.0.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.4
pandas @ file:///C:/ci/pandas_1592841747005/work
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
path.py==12.0.2
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
psutil==5.7.0
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint @ file:///C:/ci/pylint_1592487534522/work
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.7.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
regex==2020.5.14
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.6
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn @ file:///C:/ci/scikit-learn_1592847564598/work
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx @ file:///tmp/build/80754af9/sphinx_1592842202926/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport @ file:///tmp/build/80754af9/sphinxcontrib-websupport_1592842361205/work
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.17
statsmodels==0.11.1
sympy==1.6
tables==3.6.1
tblib==1.6.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
tensorflow-hub==0.5.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1592847887441/work
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.4
xlwt==1.3.0
zict==2.0.0
zipp==3.1.0
